Dtedder/402 download custom configurations fixed (#464)
* initial version for custom configuration downloads
(cherry picked/conflicts resolved)

* cleaned up downloading for portal items and plain url downloads. added authentication view for secured portal items

* missed include for QUrl

* added confirmation for removing configurations to include option to remove the entry itself from the list

* strip off any trailing slashes if they exist on the RootDataDirectory value from the downloaded package's configuration file

* when local file is the type, there is no extra download to be cleaned up. the extract is applied directly to the network path as the source. don' t delete the 'download' when extract is completed if local file is the type.

minor naming cleanup in ConfigurationListModel. added method to get configuration by name. Use full url pattern match check for portal/ago item.

* extended ZipHelper to emit a signal for every percentage extracted.

* added validation in the form inputs

* add concurrent extension to project files. updated versions for gradle for android build

* removed file cleanup that is outside of the extraction completion handler
correction for regex restricting configuration names to alphanumeric and underscore

checking in the fixes that should be applied to project files for iOS to remove the multimedia extension of Qt

* switched to a single 'progress' indicator.

* switch file open dialog to new icon

* remove the abort for cases when the head request fails to get the total download size and just give a warning message.
allow 'Delete' for items in the configurations list directly as long as the config isnt downloading or unzipping.

* cleanup from PR suggestions

* YesNoCancel option is not necessary. Regular dialog works for the functionality require for removing files only and removing the entire entry.

* use std::find_if for getting a configuration by name.

* move larger lambdas to their own functions.

* add check to prevent user from creating a new configuration with the same name.

* added logic to detect android content urls and copy to the app's downloads folder to be extracted and cleaned up like other url downloads

* check for divide by zero

rename bytesAvailable to bytesAllowed
add const qualifiers

* remove redundan calls to ->abort()
added a signal for notifying user the download failed so they can confirm removing the configuration from the list
pass the result of the file size check to the readyRead method so that it doesn't get reported in the event the request for the content also fails
prevent 'finish' slots from continuing if the finish was downstream of an actual error
add a convenience method to fetch the index of a configuration in the list by its name
connect the Options UI to the configurationDownloadFailed that was added to the ConfigurationController

* additional checks on local and web based files to make sure they can be loaded from their url. checking on first read to make sure the type is zip and not a text response from the server about a failure

* missed applying the gradle updates to Vehicle. Vehicle and Handheld files are sync'd now

* same file checks for missing local files are needed for the content:// urls for Android

* checking in formatting directly from Xcode (16.0) on plist files. adding the `LSSupportsOpeningDocumentsInPlace` key

* move location for extracting configurations to the AppDataLocation which is hidden from users through file browser on iOS

* added specific messages for the download/extract failure scenarios
added cleanup for all failures to download/extract
implemented failure if the DsaAppConfig.json file cannot be found on anything other than the default configuration from esri
moved some private non-class functions to the ConfigurationController class
added an emit error from within the ZipHelper when the file fails to open as a proper zip

* updates from PR discussion. added const to zip helper percentage method. made regex properties readonly in qml. remove ignoreSslErrors.

* switch to using names to remove entries from the configurations list